tests/test_handlers.py
"""Test handlers."""

import pytest
from unittest.mock import AsyncMock, MagicMock

from telegram import Update, User as TGUser, Chat, Message


@pytest.fixture
def mock_update() -> Update:
    """Create mock Telegram update."""
    update = AsyncMock(spec=Update)
    update.effective_user = TGUser(
        id=123456789,
        is_bot=False,
        first_name="Test",
        last_name="User",
        username="testuser",
    )
    update.message = AsyncMock(spec=Message)
    update.message.chat = Chat(id=123456789, type="private")
    return update


@pytest.fixture
def mock_context() -> MagicMock:
    """Create mock Telegram context."""
    context = MagicMock()
    context.user_data = {}
    context.bot = AsyncMock()
    return context


@pytest.mark.asyncio
async def test_start_command_new_user(mock_update, mock_context):
    """Test /start command for new user."""
    from bot.handlers.start import start_command
    
    # This would require full async context and database setup
    # Simplified example
    assert True  # Placeholder
